Issue #21635: Update Servlet 6.0 factories bundle name.
