+++ dev/com.ibm.websphere.appserver.features/visibility/private/io.openliberty.servlet.internal-6.0.feature
@@ -26,9 +26,9 @@ Subsystem-Category: JakartaEE10Application
  com.ibm.ws.http.plugin.merge, \
  com.ibm.ws.webserver.plugin.runtime.jakarta, \
  com.ibm.ws.webserver.plugin.runtime.interfaces,\
- io.openliberty.servlet.6.0.internal.factories, \
  io.openliberty.session.6.0.internal, \
- io.openliberty.webcontainer.servlet.6.0.internal
+ io.openliberty.webcontainer.servlet.6.0.internal, \
+ io.openliberty.webcontainer.servlet.6.0.internal.factories
 -jars=com.ibm.ws.webserver.plugin.utility
 -files=bin/tools/ws-webserverPluginutil.jar, \
  bin/pluginUtility; ibm.executable:=true; ibm.file.encoding:=ebcdic, \
