--- dev/io.openliberty.servlet.6.0.internal.factories/bnd.bnd
@@ -1,103 +1,0 @@
-#*******************************************************************************
-# Copyright (c) 2022 IBM Corporation and others.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License v1.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-v10.html
-#
-# Contributors:
-#     IBM Corporation - initial API and implementation
-#*******************************************************************************
--include= ~../cnf/resources/bnd/bundle.props
-bVersion=1.0
-
-javac.source: 11
-javac.target: 11
-
-Require-Capability: osgi.ee; filter:="(&(osgi.ee=JavaSE)(version=11))"
-
-Bundle-Name: WAS WebContainer
-Bundle-SymbolicName: io.openliberty.servlet.6.0.internal.factories
-Bundle-Description: WAS WebContainer with object factories for Servlet 6.0 support
-WLP-ServerName: IBM WebSphere Application Server
-WLP-ServerVersion: ${liberty.beta.version}
-
--dsannotations: \
-  io.openliberty.servlet.internal.async.factory.AsyncContextFactory60Impl, \
-  io.openliberty.servlet.internal.osgi.container.config.factory.WebAppConfiguratorFactory60Impl, \
-  io.openliberty.servlet.internal.osgi.response.factory.IRequestResponseFactory60Impl, \
-  io.openliberty.servlet.internal.osgi.srt.factory.SRTConnectionContextPool60Impl, \
-  io.openliberty.servlet.internal.osgi.webapp.factory.WebAppFactory60Impl, \
-  io.openliberty.servlet.internal.servlet.factory.CacheServletWrapperFactory60Impl, \
-  io.openliberty.servlet.internal.session.impl.factory.SessionContextRegistryImplFactory60Impl, \
-  io.openliberty.servlet.internal.util.factory.URIMatcherFactory60Impl
-
-Service-Component: \
-  com.ibm.ws.webcontainer.v60.dd; \
-    implementation:=com.ibm.ws.javaee.version.ServletVersion; \
-    provide:=com.ibm.ws.javaee.version.ServletVersion; \
-    properties:="version:Integer=60", \
-  com.ibm.ws.webcontainer.v60.httpTransportBehavior; \
-    implementation:=com.ibm.wsspi.http.ee7.HttpTransportBehavior; \
-    provide:=com.ibm.wsspi.http.ee7.HttpTransportBehavior; \
-    properties:="useEE7Streams:Boolean=true,useIOEBehavior:Boolean=true", \
-  com.ibm.ws.webcontainer.v60.httpProtocolBehavior; \
-  	implementation:=com.ibm.wsspi.channelfw.HttpProtocolBehavior; \
-  	provide:=com.ibm.wsspi.channelfw.HttpProtocolBehavior; \
-  	properties:="httpVersionSetting:String=2.0_Optional_On"
-
-# For each exported package, create (in that package) a package-info.java
-# file, and place an @version javadoc tag in package-level javadoc. 
-# Append ";provide:=true" if this bundle also provides an implementation
-# for the exported API.
-Export-Package: \
-    com.ibm.ws.webcontainer.speclevel, \
-    io.openliberty.servlet.internal.*factory*;provide:=true
-
-Import-Package: \
-    !io.openliberty.servlet.internal.async.factory, \
-    !io.openliberty.servlet.internal.osgi.container.config.factory, \
-    !io.openliberty.servlet.internal.osgi.response.factory, \
-    !io.openliberty.servlet.internal.osgi.srt.factory, \
-    !io.openliberty.servlet.internal.osgi.webapp.factory, \
-    !io.openliberty.servlet.internal.servlet.factory, \
-    !io.openliberty.servlet.internal.session.impl.factory, \
-    !io.openliberty.servlet.internal.util.factory, \
-    com.ibm.websphere.csi, \
-    com.ibm.ws.webcontainer40.osgi.container.config, \
-    com.ibm.ws.webcontainer40.osgi.request, \
-    com.ibm.ws.webcontainer40.osgi.srt, \
-    com.ibm.ws.webcontainer31.osgi.response, \
-    com.ibm.ws.webcontainer31.osgi.srt, \
-    com.ibm.ws.webcontainer31.async, \
-    com.ibm.ws.webcontainer31.osgi.webapp, \
-    com.ibm.ws.webcontainer31.session.impl, \
-    com.ibm.ws.webcontainer.osgi.srt, \
-    com.ibm.ws.container.service.metadata, \
-    com.ibm.ws.webcontainer.osgi.webapp, \
-    com.ibm.wsspi.injectionengine, \
-    io.openliberty.webcontainer60.session.impl, \
-    io.openliberty.webcontainer60.osgi.request, \
-    io.openliberty.webcontainer60.osgi.srt, \
-    *
-
--buildpath: \
-    com.ibm.ws.webcontainer.servlet.3.1.jakarta;version=latest,\
-    com.ibm.ws.webcontainer.jakarta;version=latest,\
-    com.ibm.ws.logging.core,\
-    com.ibm.ws.adaptable.module;version=latest,\
-    com.ibm.websphere.org.osgi.core;version=latest,\
-    com.ibm.websphere.org.osgi.service.component;version=latest,\
-    com.ibm.wsspi.org.osgi.service.component.annotations;version=latest,\
-    com.ibm.ws.transport.http;version=latest,\
-    com.ibm.ws.channelfw;version=latest,\
-    com.ibm.ws.managedobject;version=latest,\
-    com.ibm.ws.container.service;version=latest,\
-    com.ibm.ws.injection;version=latest,\
-    com.ibm.ws.session.jakarta;version=latest,\
-    com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
-    com.ibm.ws.webcontainer.servlet.4.0.jakarta;version=latest,\
-    com.ibm.ws.resource;version=latest, \
-	com.ibm.ws.kernel.boot.core;version=latest,\
-    io.openliberty.jakarta.servlet.6.0;version=latest,\
-    io.openliberty.webcontainer.servlet.6.0.internal;version=latest
